INDEV 0.1 - Big Fixes
Fixed the bug in which the word was not equal to the "guessed" word.

Problem was that when grabbing the words from the list, I was using readlines(), which adds the new line escape sequence at the end of each line. It made it LOOK like the words were the same, however they weren't for this reason.